==> 02_script/01-initial_analysis.qmd <==
---
title: "Exploratory analysis of reconvictions data"
author: "Ben Matthews"
format: gfm
editor: visual
---

## Setting up

Load up packages.

```{r}
library(readxl)
library(here)
library(tidyverse)
```

Read in the data. There are lots of sheets - this has age and gender data for each 'LA' from 2004-21.

```{r, warning=FALSE}
tmp <- read_xlsx(here("01_data",
               "reconvictions-2020-21-offender-cohort-additional-datasets.xlsx"),
          sheet = "AGdata",
          skip = 4)

tmp <- 
tmp |> 
  select(-...11:-...14) |> 
  select(-...1) |> 
  rename(la = ...2,
         year = ...3) |> 
  janitor::clean_names() |> 
  filter(la != "LA")

```

## Initial analysis

Filter out the aggregated rows and then just plot the raw numbers of people convicted across LAs in each year by age and gender

```{r, fig.width=12,fig.height=10}
reconv <- 
tmp |> 
  filter(age != "All",
         gender != "All",
         la != "Not known")

reconv |> 
  ggplot(aes(x = year, y = number_of_offenders, colour = age, group = interaction(age, gender))) +
  geom_line() +
  facet_wrap(la ~ gender, scales = "free_y")

```

So it looks like the trend is generally downwards for young people but less clear for older people, regardless of gender. We can test whether the linear trend for each age is downwards pretty simply. I'm not sure if we need the *quasi* poisson bit here but feels like it should help with the confidence intervals and stuff.

```{r, fig.width=12,fig.height=10}
reconv <- 
reconv |> 
  mutate(year = as.numeric(str_sub(year, 1, 4)))

reconv <- 
reconv |> 
  mutate(age = factor(age,
                      levels = c("under 21",
                                 "21 to 25",
                                 "26 to 30",
                                 "31 to 40",
                                 "over 40")))

reconv |> 
  group_by(la, gender, age) |> 
  nest() |> 
  mutate(model = map(data, ~ glm(number_of_offenders ~ year, family = "quasipoisson", data = .x)),
         results = map(model, broom::tidy)) |> 
  unnest(results) |> 
  filter(term!= "(Intercept)") |> 
  mutate(conf_low = estimate - 1.96 * std.error,
         conf_upp = estimate + 1.96 * std.error) |> 
  ggplot(aes(x = estimate, y = la, colour = age)) +
  geom_vline(xintercept = 0) +
  geom_point(position = position_dodge(width = -0.5)) +
  geom_errorbar(aes(xmin = conf_low, xmax = conf_upp),
                position = position_dodge(width = -0.5)) +
  facet_wrap(~ gender)


```

For every LA (and high court) the linear slope is negative for young people, and for most of the 21-25 group too. For most of the LAs there is no linear trend for older people.

So the demographics of people convicted in Scotland are changing quite a lot. Implications: all else equal, more people being convicted will be parents (as they are more likely to be of parenting age)? Could check the overall parity statistics by age to assess this.

## Something simpler

Maybe the easiest way to do this is just look at the proportion of people overall convicted by age.

```{r}

agg_dat <- 
tmp |> 
  filter(gender != "All",
         age != "All") |> 
  group_by(year, gender, age) |> 
  summarise(n = sum(number_of_offenders)) |> 
  group_by(gender, year) |> 
  mutate(prop = n / sum(n)) |> 
  ungroup() |> 
   mutate(age = factor(age,
                      levels = c("under 21",
                                 "21 to 25",
                                 "26 to 30",
                                 "31 to 40",
                                 "over 40")))

agg_dat |> 
  ggplot(aes(x = year, y = prop, fill = fct_rev(age), colour = fct_rev(age))) +
  geom_bar(stat = "identity") +
  facet_grid(~ gender) 
```
